requirements.txt
Pillow
tensorflow==2.6.2
flask